requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.25
python-dotenv==1.0.0
pytest==6.2.5
supabase==1.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.0.post1
python-jose==3.3.0
Flask-Cors==4.0.0
gunicorn==21.2.0
